requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6
pymongo==4.6.1
python-dotenv==1.0.0
bcrypt==4.1.2
Werkzeug==3.0.1
PyPDF2==3.0.1
email-validator==2.1.0
Pillow==10.2.0
pandas==2.2.0
openpyxl==3.1.2
requests==2.31.0
cloudinary==1.37.0
gunicorn==21.2.0
python-socketio==5.11.0
eventlet==0.33.3
google-generativeai==0.3.2
pytesseract==0.3.10